Made New Architecture with MockConnection
