+++ my-no-sql-tcp-reader/src/get_entities_builder/get_entities_builder_inner.rs
@@ -4,22 +4,28 @@ use my_no_sql_abstractions::MyNoSqlEntity;
 use my_no_sql_tcp_shared::sync_to_main::UpdateEntityStatisticsData;
 use rust_extensions::date_time::DateTimeAsMicroseconds;
 
-use super::super::my_no_sql_data_reader_tcp::MyNoSqlDataReaderInner;
+use crate::{my_no_sql_connector::MyNoSqlConnector, MyNoSqlDataReaderInner};
 
 pub struct GetEntitiesBuilderInner<TMyNoSqlEntity: MyNoSqlEntity + Sync + Send + 'static> {
     partition_key: String,
     update_statistic_data: UpdateEntityStatisticsData,
     inner: Arc<MyNoSqlDataReaderInner<TMyNoSqlEntity>>,
+    connector: Arc<dyn MyNoSqlConnector + Send + Sync + 'static>,
 }
 
 impl<TMyNoSqlEntity: MyNoSqlEntity + Sync + Send + 'static>
     GetEntitiesBuilderInner<TMyNoSqlEntity>
 {
-    pub fn new(partition_key: String, inner: Arc<MyNoSqlDataReaderInner<TMyNoSqlEntity>>) -> Self {
+    pub fn new(
+        partition_key: String,
+        inner: Arc<MyNoSqlDataReaderInner<TMyNoSqlEntity>>,
+        connector: Arc<dyn MyNoSqlConnector + Send + Sync + 'static>,
+    ) -> Self {
         Self {
             partition_key,
             update_statistic_data: UpdateEntityStatisticsData::default(),
             inner,
+            connector,
         }
     }
 
@@ -41,19 +47,20 @@ impl<TMyNoSqlEntity: MyNoSqlEntity + Sync + Send + 'static>
 
     pub async fn get_as_vec(&self) -> Option<Vec<Arc<TMyNoSqlEntity>>> {
         let db_rows = {
-            let reader = self.inner.get_data().read().await;
+            let reader = self.inner.data.read().await;
             reader.get_by_partition_as_vec(self.partition_key.as_str())
         }?;
 
-        self.inner
-            .get_sync_handler()
-            .update(
-                TMyNoSqlEntity::TABLE_NAME,
-                &self.partition_key,
-                || db_rows.iter().map(|itm| itm.get_row_key()),
-                &self.update_statistic_data,
-            )
-            .await;
+        if let Some(sync_handler) = self.connector.get_sync_handler() {
+            sync_handler
+                .update(
+                    TMyNoSqlEntity::TABLE_NAME,
+                    &self.partition_key,
+                    || db_rows.iter().map(|itm| itm.get_row_key()),
+                    &self.update_statistic_data,
+                )
+                .await;
+        }
 
         Some(db_rows)
     }
@@ -63,38 +70,40 @@ impl<TMyNoSqlEntity: MyNoSqlEntity + Sync + Send + 'static>
         filter: impl Fn(&TMyNoSqlEntity) -> bool,
     ) -> Option<Vec<Arc<TMyNoSqlEntity>>> {
         let db_rows = {
-            let reader = self.inner.get_data().read().await;
+            let reader = self.inner.data.read().await;
             reader.get_by_partition_as_vec_with_filter(&self.partition_key, filter)
         }?;
 
-        self.inner
-            .get_sync_handler()
-            .update(
-                TMyNoSqlEntity::TABLE_NAME,
-                &self.partition_key,
-                || db_rows.iter().map(|itm| itm.get_row_key()),
-                &self.update_statistic_data,
-            )
-            .await;
+        if let Some(sync_handler) = self.connector.get_sync_handler() {
+            sync_handler
+                .update(
+                    TMyNoSqlEntity::TABLE_NAME,
+                    &self.partition_key,
+                    || db_rows.iter().map(|itm| itm.get_row_key()),
+                    &self.update_statistic_data,
+                )
+                .await;
+        }
 
         Some(db_rows)
     }
 
     pub async fn get_as_btree_map(&self) -> Option<BTreeMap<String, Arc<TMyNoSqlEntity>>> {
         let db_rows = {
-            let reader = self.inner.get_data().read().await;
-            reader.get_by_partition(&self.partition_key)
+            let reader = self.inner.data.read().await;
+            reader.get_by_partition(&self.partition_key).cloned()
         }?;
 
-        self.inner
-            .get_sync_handler()
-            .update(
-                TMyNoSqlEntity::TABLE_NAME,
-                &self.partition_key,
-                || db_rows.values().map(|itm| itm.get_row_key()),
-                &self.update_statistic_data,
-            )
-            .await;
+        if let Some(sync_handler) = self.connector.get_sync_handler() {
+            sync_handler
+                .update(
+                    TMyNoSqlEntity::TABLE_NAME,
+                    &self.partition_key,
+                    || db_rows.values().map(|itm| itm.get_row_key()),
+                    &self.update_statistic_data,
+                )
+                .await;
+        }
 
         Some(db_rows)
     }
@@ -104,19 +113,20 @@ impl<TMyNoSqlEntity: MyNoSqlEntity + Sync + Send + 'static>
         filter: impl Fn(&TMyNoSqlEntity) -> bool,
     ) -> Option<BTreeMap<String, Arc<TMyNoSqlEntity>>> {
         let db_rows = {
-            let reader = self.inner.get_data().read().await;
+            let reader = self.inner.data.read().await;
             reader.get_by_partition_with_filter(&self.partition_key, filter)
         }?;
 
-        self.inner
-            .get_sync_handler()
-            .update(
-                TMyNoSqlEntity::TABLE_NAME,
-                &self.partition_key,
-                || db_rows.values().map(|itm| itm.get_row_key()),
-                &self.update_statistic_data,
-            )
-            .await;
+        if let Some(sync_handler) = self.connector.get_sync_handler() {
+            sync_handler
+                .update(
+                    TMyNoSqlEntity::TABLE_NAME,
+                    &self.partition_key,
+                    || db_rows.values().map(|itm| itm.get_row_key()),
+                    &self.update_statistic_data,
+                )
+                .await;
+        }
 
         Some(db_rows)
     }
