+++ my-no-sql-tcp-reader/src/get_entities_builder/mod.rs
@@ -2,7 +2,3 @@ mod get_entities_builder;
 pub use get_entities_builder::*;
 mod get_entities_builder_inner;
 pub use get_entities_builder_inner::*;
-#[cfg(feature = "mocks")]
-mod get_entities_builder_mock;
-#[cfg(feature = "mocks")]
-pub use get_entities_builder_mock::*;
