+++ my-no-sql-tcp-reader/src/get_entity_builder/get_entity_builder_inner.rs
@@ -4,13 +4,14 @@ use my_no_sql_abstractions::MyNoSqlEntity;
 use my_no_sql_tcp_shared::sync_to_main::UpdateEntityStatisticsData;
 use rust_extensions::date_time::DateTimeAsMicroseconds;
 
-use super::super::my_no_sql_data_reader_tcp::MyNoSqlDataReaderInner;
+use crate::{my_no_sql_connector::MyNoSqlConnector, MyNoSqlDataReaderInner};
 
 pub struct GetEntityBuilderInner<'s, TMyNoSqlEntity: MyNoSqlEntity + Sync + Send + 'static> {
     partition_key: &'s str,
     row_key: &'s str,
     update_statistic_data: UpdateEntityStatisticsData,
     inner: Arc<MyNoSqlDataReaderInner<TMyNoSqlEntity>>,
+    connector: Arc<dyn MyNoSqlConnector + Send + Sync + 'static>,
 }
 
 impl<'s, TMyNoSqlEntity: MyNoSqlEntity + Sync + Send + 'static>
@@ -20,12 +21,14 @@ impl<'s, TMyNoSqlEntity: MyNoSqlEntity + Sync + Send + 'static>
         partition_key: &'s str,
         row_key: &'s str,
         inner: Arc<MyNoSqlDataReaderInner<TMyNoSqlEntity>>,
+        connector: Arc<dyn MyNoSqlConnector + Send + Sync + 'static>,
     ) -> Self {
         Self {
             partition_key,
             row_key,
             update_statistic_data: UpdateEntityStatisticsData::default(),
             inner,
+            connector,
         }
     }
 
@@ -47,20 +50,21 @@ impl<'s, TMyNoSqlEntity: MyNoSqlEntity + Sync + Send + 'static>
 
     pub async fn execute(&self) -> Option<Arc<TMyNoSqlEntity>> {
         let result = {
-            let reader = self.inner.get_data().read().await;
+            let reader = self.inner.data.read().await;
             reader.get_entity(self.partition_key, self.row_key)
         };
 
         if result.is_some() {
-            self.inner
-                .get_sync_handler()
-                .update(
-                    TMyNoSqlEntity::TABLE_NAME,
-                    self.partition_key,
-                    || [self.row_key].into_iter(),
-                    &self.update_statistic_data,
-                )
-                .await;
+            if let Some(sync_handler) = self.connector.get_sync_handler() {
+                sync_handler
+                    .update(
+                        TMyNoSqlEntity::TABLE_NAME,
+                        self.partition_key,
+                        || [self.row_key].into_iter(),
+                        &self.update_statistic_data,
+                    )
+                    .await;
+            }
         }
 
         result
