+++ my-no-sql-tcp-reader/src/lib.rs
@@ -1,11 +1,27 @@
-mod my_no_sql_tcp_connection;
+mod callback_triggers;
+mod my_no_sql_connector;
+mod my_no_sql_data_reader;
+mod my_no_sql_data_reader_callbacks;
+mod my_no_sql_data_reader_callbacks_pusher;
+mod my_no_sql_data_reader_data;
+mod my_no_sql_data_reader_inner;
 mod settings;
 mod subscribers;
-mod tcp_events;
+pub use my_no_sql_data_reader_data::MyNoSqlDataReaderData;
 
-pub use my_no_sql_tcp_connection::MyNoSqlTcpConnection;
+pub use my_no_sql_data_reader_callbacks::MyNoSqlDataReaderCallBacks;
+pub use my_no_sql_data_reader_callbacks_pusher::MyNoSqlDataReaderCallBacksPusher;
+pub mod tcp;
+
+mod get_entities_builder;
+mod get_entity_builder;
+
+pub use get_entities_builder::*;
+pub use get_entity_builder::*;
+
+pub use my_no_sql_data_reader::*;
+pub use my_no_sql_data_reader_inner::*;
 pub use settings::*;
-pub use subscribers::{MyNoSqlDataReaderCallBacks, MyNoSqlDataReaderData, MyNoSqlDataReaderTcp, MyNoSqlDataReader};
 
 #[cfg(feature = "mocks")]
-pub use subscribers::MyNoSqlDataReaderMock;
+pub mod mock;
