+++ my-no-sql-tcp-reader/src/my_no_sql_connector.rs
@@ -1,0 +1,5 @@
+use my_no_sql_tcp_shared::sync_to_main::SyncToMainNodeHandler;
+
+pub trait MyNoSqlConnector {
+    fn get_sync_handler(&self) -> Option<&SyncToMainNodeHandler>;
+}
