+++ my-no-sql-tcp-reader/src/subscribers/mod.rs
@@ -1,28 +1,4 @@
-mod callback_triggers;
-mod get_entities_builder;
-mod get_entity_builder;
-mod my_no_sql_data_reader;
-mod my_no_sql_data_reader_callbacks;
-mod my_no_sql_data_reader_callbacks_pusher;
-mod my_no_sql_data_reader_data;
-mod my_no_sql_data_reader_tcp;
 mod subscribers;
-mod update_event_trait;
-pub use my_no_sql_data_reader_data::MyNoSqlDataReaderData;
-pub use my_no_sql_data_reader_tcp::MyNoSqlDataReaderTcp;
-
-pub use get_entities_builder::*;
-pub use get_entity_builder::*;
-pub use my_no_sql_data_reader::*;
-pub use my_no_sql_data_reader_callbacks::MyNoSqlDataReaderCallBacks;
-pub use my_no_sql_data_reader_callbacks_pusher::MyNoSqlDataReaderCallBacksPusher;
-pub use subscribers::Subscribers;
-pub use update_event_trait::UpdateEvent;
-#[cfg(feature = "mocks")]
-mod my_no_sql_data_reader_mock;
-#[cfg(feature = "mocks")]
-pub use my_no_sql_data_reader_mock::*;
-#[cfg(feature = "mocks")]
-mod my_no_sql_data_reader_mock_inner;
-#[cfg(feature = "mocks")]
-pub use my_no_sql_data_reader_mock_inner::*;
+pub use subscribers::*;
+mod data_reader_update;
+pub use data_reader_update::*;
