+++ my-no-sql-tcp-reader/src/subscribers/subscribers.rs
@@ -1,14 +1,10 @@
 use std::{collections::BTreeMap, sync::Arc};
 
-use my_no_sql_abstractions::{MyNoSqlEntity, MyNoSqlEntitySerializer};
-use my_no_sql_tcp_shared::sync_to_main::SyncToMainNodeHandler;
-use rust_extensions::ApplicationStates;
+use super::DataReaderUpdater;
 use tokio::sync::RwLock;
 
-use super::{MyNoSqlDataReaderTcp, UpdateEvent};
-
 pub struct Subscribers {
-    subscribers: RwLock<BTreeMap<String, Arc<dyn UpdateEvent + Send + Sync + 'static>>>,
+    subscribers: RwLock<BTreeMap<&'static str, Arc<dyn DataReaderUpdater + Send + Sync + 'static>>>,
 }
 
 impl Subscribers {
@@ -18,36 +14,24 @@ impl Subscribers {
         }
     }
 
-    pub async fn create_subscriber<TMyNoSqlEntity>(
+    pub async fn add(
         &self,
-        app_states: Arc<dyn ApplicationStates + Send + Sync + 'static>,
-        sync_handler: Arc<SyncToMainNodeHandler>,
-    ) -> Arc<MyNoSqlDataReaderTcp<TMyNoSqlEntity>>
-    where
-        TMyNoSqlEntity: MyNoSqlEntity + MyNoSqlEntitySerializer + Sync + Send + 'static,
-    {
+        table_name: &'static str,
+        subscriber: Arc<dyn DataReaderUpdater + Send + Sync + 'static>,
+    ) {
         let mut write_access = self.subscribers.write().await;
 
-        if write_access.contains_key(TMyNoSqlEntity::TABLE_NAME) {
-            panic!(
-                "You already subscribed for the table {}",
-                TMyNoSqlEntity::TABLE_NAME
-            );
+        if write_access.contains_key(table_name) {
+            panic!("You already subscribed for the table {}", table_name);
         }
 
-        let new_reader = MyNoSqlDataReaderTcp::new(app_states, sync_handler).await;
-
-        let new_reader = Arc::new(new_reader);
-
-        write_access.insert(TMyNoSqlEntity::TABLE_NAME.to_string(), new_reader.clone());
-
-        new_reader
+        write_access.insert(table_name, subscriber);
     }
 
     pub async fn get(
         &self,
         table_name: &str,
-    ) -> Option<Arc<dyn UpdateEvent + Send + Sync + 'static>> {
+    ) -> Option<Arc<dyn DataReaderUpdater + Send + Sync + 'static>> {
         let read_access = self.subscribers.write().await;
         let result = read_access.get(table_name)?;
         Some(result.clone())
